+++ src/main/java/jak/dut/Investment/model/calculation/CalculaltionAlgorithmStrategy.java
@@ -1,8 +1,10 @@
 package jak.dut.Investment.model.calculation;
 
+import java.math.BigDecimal;
+
 @FunctionalInterface
 public interface CalculaltionAlgorithmStrategy {
 
-    Double calculation(Calculation calculation);
+    BigDecimal calculation(Calculation calculation);
 
 }
